requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-postgresql==5.0.0
